#94: Move unit test to CategorySyncTest.
