+++ src/test/java/com/commercetools/sync/categories/CategorySyncTest.java
@@ -27,6 +27,9 @@ import static java.lang.String.format;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 public class CategorySyncTest {
@@ -420,5 +423,88 @@ public class CategorySyncTest {
             .build();
         final boolean doesRequire = CategorySync.requiresChangeParentUpdateAction(category, categoryDraft);
         assertThat(doesRequire).isFalse();
+    }
+
+    @Test
+    public void sync_WithBatchSizeSet_ShouldCallSyncOnEachBatch() {
+        final int batchSize = 1;
+        CategorySyncOptions categorySyncOptions = CategorySyncOptionsBuilder
+            .of(mock(SphereClient.class))
+            .setErrorCallBack(
+                (errorMessage, exception) -> {
+                    errorCallBackMessages.add(errorMessage);
+                    errorCallBackExceptions.add(exception);
+                })
+            .setBatchSize(batchSize)
+            .build();
+
+        final CategorySync categorySync =
+            new CategorySync(categorySyncOptions, getMockTypeService(), getMockCategoryService());
+
+        final ArrayList<CategoryDraft> categoryDrafts = new ArrayList<>();
+
+        final int numberOfCategoryDrafts = 160;
+        for (int i = 0; i < numberOfCategoryDrafts; i++) {
+            categoryDrafts.add(getMockCategoryDraft(Locale.ENGLISH, "name", "key" + i, "parentKey",
+                "customTypeId", new HashMap<>()));
+        }
+
+        final CategorySync mockCategorySync = spy(categorySync);
+
+        final CategorySyncStatistics syncStatistics = mockCategorySync.sync(categoryDrafts)
+                                                                      .toCompletableFuture().join();
+
+        int expectedNumberOfCalls = (int) (Math.ceil(numberOfCategoryDrafts / batchSize) + 1);
+        verify(mockCategorySync, times(expectedNumberOfCalls)).syncBatches(any(), any());
+
+        int expectedNumberOfCategoriesCreated = expectedNumberOfCalls - 1;
+        assertThat(syncStatistics.getCreated()).isEqualTo(expectedNumberOfCategoriesCreated);
+        assertThat(syncStatistics.getFailed()).isEqualTo(0);
+        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
+        assertThat(syncStatistics.getProcessed()).isEqualTo(categoryDrafts.size());
+        assertThat(syncStatistics.getReportMessage()).isEqualTo(format("Summary: %s categories were processed"
+                + " in total (%s created, 0 updated, 0 failed to sync and %s categories with a missing parent).",
+            categoryDrafts.size(),
+            expectedNumberOfCategoriesCreated,
+            categoryDrafts.size()));
+        assertThat(errorCallBackMessages).hasSize(0);
+        assertThat(errorCallBackExceptions).hasSize(0);
+
+        // With Default batch size
+
+        categorySyncOptions = CategorySyncOptionsBuilder.of(mock(SphereClient.class))
+                                                        .setErrorCallBack((errorMessage, exception) -> {
+                                                            errorCallBackMessages.add(errorMessage);
+                                                            errorCallBackExceptions.add(exception);
+                                                        })
+                                                        .build();
+
+        final CategorySync categorySyncWithDefaultBatchSize = new CategorySync(categorySyncOptions,
+            getMockTypeService(), getMockCategoryService());
+        final CategorySync mockCategorySyncWithDefaultBatchSize = spy(categorySyncWithDefaultBatchSize);
+
+        final CategorySyncStatistics syncStatisticsWithDefaultBatchSize = mockCategorySyncWithDefaultBatchSize
+            .sync(categoryDrafts)
+            .toCompletableFuture().join();
+
+        expectedNumberOfCalls =
+            (int) (Math.ceil(numberOfCategoryDrafts / (double) CategorySyncOptionsBuilder.BATCH_SIZE_DEFAULT) + 1);
+
+        verify(mockCategorySyncWithDefaultBatchSize, times(expectedNumberOfCalls)).syncBatches(any(), any());
+
+
+        expectedNumberOfCategoriesCreated = expectedNumberOfCalls - 1;
+        final int expectedFailedCategories = categoryDrafts.size() - (expectedNumberOfCategoriesCreated);
+
+        assertThat(syncStatisticsWithDefaultBatchSize.getCreated()).isEqualTo(expectedNumberOfCategoriesCreated);
+        assertThat(syncStatisticsWithDefaultBatchSize.getFailed()).isEqualTo(expectedFailedCategories);
+        assertThat(syncStatisticsWithDefaultBatchSize.getUpdated()).isEqualTo(0);
+        assertThat(syncStatisticsWithDefaultBatchSize.getProcessed()).isEqualTo(categoryDrafts.size());
+        assertThat(syncStatisticsWithDefaultBatchSize.getReportMessage())
+            .isEqualTo(format("Summary: %s categories were processed in total (%s created, 0 updated, %s failed to sync"
+                    + " and %s categories with a missing parent).", categoryDrafts.size(),
+                expectedNumberOfCategoriesCreated, expectedFailedCategories, categoryDrafts.size()));
+        assertThat(errorCallBackMessages).hasSize(0);
+        assertThat(errorCallBackExceptions).hasSize(0);
     }
 }
